+++ data/src/main/kotlin/com/vultisig/wallet/data/models/Coins.kt
@@ -1,0 +1,1586 @@
+package com.vultisig.wallet.data.models
+
+import com.vultisig.wallet.data.models.Chain.BscChain
+import com.vultisig.wallet.data.models.Chain.Terra
+import com.vultisig.wallet.data.models.Chain.TerraClassic
+
+object Coins {
+
+    val wewe = Coin(
+        chain = Chain.Base,
+        ticker = "WEWE",
+        logo = "wewe",
+        address = "",
+        decimal = 18,
+        hexPublicKey = "",
+        priceProviderID = "",
+        contractAddress = "0x6b9bb36519538e0C073894E964E90172E1c0B41F",
+        isNativeToken = false,
+    )
+    val solana = Coin(
+        chain = Chain.Solana,
+        ticker = "SOL",
+        logo = "sol",
+        address = "",
+        decimal = 9,
+        hexPublicKey = "",
+        priceProviderID = "solana",
+        contractAddress = "",
+        isNativeToken = true,
+    )
+    val polkadot = Coin(
+        chain = Chain.Polkadot,
+        ticker = "DOT",
+        logo = "dot",
+        address = "",
+        decimal = 10,
+        hexPublicKey = "",
+        priceProviderID = "polkadot",
+        contractAddress = "",
+        isNativeToken = true,
+    )
+
+    val SupportedCoins = listOf(
+        wewe,
+        Coin(
+            chain = Chain.Bitcoin,
+            ticker = "BTC",
+            logo = "btc",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "bitcoin",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.BitcoinCash,
+            ticker = "BCH",
+            logo = "bch",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "bitcoin-cash",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Litecoin,
+            ticker = "LTC",
+            logo = "ltc",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "litecoin",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Dogecoin,
+            ticker = "DOGE",
+            logo = "doge",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "dogecoin",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Dash,
+            ticker = "DASH",
+            logo = "dash",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "dash",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.ThorChain,
+            ticker = "RUNE",
+            logo = "rune",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "thorchain",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.MayaChain,
+            ticker = "CACAO",
+            logo = "cacao",
+            address = "",
+            decimal = 10,
+            hexPublicKey = "",
+            priceProviderID = "cacao",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.MayaChain,
+            ticker = "MAYA",
+            logo = "maya",
+            address = "",
+            decimal = 4,
+            hexPublicKey = "",
+            priceProviderID = "maya",
+            contractAddress = "",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "ETH",
+            logo = "eth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "usd-coin",
+            contractAddress = "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "USDT",
+            logo = "usdt",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "tether",
+            contractAddress = "0xdAC17F958D2ee523a2206206994597C13D831ec7",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "UNI",
+            logo = "uni",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "uniswap",
+            contractAddress = "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "POL",
+            logo = "pol",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "matic-network",
+            contractAddress = "0x7d1afa7b718fb893db30a3abc0cfc608aacfebb0",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "WBTC",
+            logo = "wbtc",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "wrapped-bitcoin",
+            contractAddress = "0x2260fac5e5542a773aa44fbcfedf7c193bc2c599",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "LINK",
+            logo = "link",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "chainlink",
+            contractAddress = "0x514910771af9ca656af840dff83e8264ecf986ca",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "FLIP",
+            logo = "flip",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "chainflip",
+            contractAddress = "0x826180541412d574cf1336d22c0c0a287822678a",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "TGT",
+            logo = "tgt",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "thorwallet",
+            contractAddress = "0x108a850856Db3f85d0269a2693D896B394C80325",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "FOX",
+            logo = "fox",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "shapeshift-fox-token",
+            contractAddress = "0xc770eefad204b5180df6a14ee197d99d808ee52d",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "DAI",
+            logo = "dai",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "dai",
+            contractAddress = "0x6b175474e89094c44da98b954eedeac495271d0f",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "WETH",
+            logo = "weth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "weth",
+            contractAddress = "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "YFI",
+            logo = "yfi",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "yearn-finance",
+            contractAddress = "0x0bc529c00c6401aef6d220be8c6ea1667f6ad93e",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "AAVE",
+            logo = "aave",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "aave",
+            contractAddress = "0x7fc66500c84a76ad7e9c93437bfc5ac33e2ddae9",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "COMP",
+            logo = "comp",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "compound-governance-token",
+            contractAddress = "0xc00e94cb662c3520282e6f5717214004a7f26888",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "BAT",
+            logo = "bat",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "basic-attention-token",
+            contractAddress = "0x0d8775f648430679a709e98d2b0cb6250d2887ef",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "SNX",
+            logo = "snx",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "havven",
+            contractAddress = "0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "BAL",
+            logo = "bal",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "balancer",
+            contractAddress = "0xba100000625a3754423978a60c9317c58a424e3d",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "SUSHI",
+            logo = "sushi",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "sushi",
+            contractAddress = "0x6b3595068778dd592e39a122f4f5a5cf09c90fe2",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "MKR",
+            logo = "mkr",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "maker",
+            contractAddress = "0x9f8f72aa9304c8b593d555f12ef6589cc3a579a2",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "KNC",
+            logo = "knc",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "kyber-network-crystal",
+            contractAddress = "0xdefa4e8a7bcba345f687a2f1456f5edd9ce97202",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "GRT",
+            logo = "grt",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "the-graph",
+            contractAddress = "0xc944e90c64b2c07662a292be6244bdf05cda44a7",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Ethereum,
+            ticker = "PEPE",
+            logo = "pepe",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "pepe",
+            contractAddress = "0x6982508145454ce325ddbe47a25d4ec3d2311933",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "AVAX",
+            logo = "avax",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "avalanche-2",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "usd-coin",
+            contractAddress = "0xb97ef9ef8734c71904d8002f8b6bc66dd9c48a6e",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "USDT",
+            logo = "usdt",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x9702230A8Ea53601f5cD2dc00fDBc13d4dF4A8c7",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "BTC.b",
+            logo = "BTC.b",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x152b9d0FdC40C096757F570A51E494bd4b943E50",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "sAVAX",
+            logo = "savax",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x2b2C81e08f1Af8835a78Bb2A90AE924ACE0eA4bE",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "JOE",
+            logo = "joe",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x6e84a6216eA6dACC71eE8E6b0a5B7322EEbC0fDd",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "PNG",
+            logo = "png",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x60781C2586D68229fde47564546784ab3fACA982",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "WAVAX",
+            logo = "wavax",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "aAvaUSDC",
+            logo = "aAvaUSDC",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x625E7708f30cA75bfd92586e17077590C60eb4cD",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "BLS",
+            logo = "bls",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x46B9144771Cb3195D66e4EDA643a7493fADCAF9D",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Avalanche,
+            ticker = "COQ",
+            logo = "coq",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x420FcA0121DC28039145009570975747295f2329",
+            isNativeToken = false,
+        ),
+        solana,
+        Coin(
+            chain = Chain.Solana,
+            ticker = "JUP",
+            logo = "jup",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "jupiter-exchange-solana",
+            contractAddress = "JUPyiwrYJFskUPiHa7hkeR8VUtAeFoSYbKedZNsDvCN",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Solana,
+            ticker = "WIF",
+            logo = "wif",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "dogwifcoin",
+            contractAddress = "EKpQGSJtjMFqKZ9KQanSqYXRcF8fBopzLHYxdM65zcjm",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Solana,
+            ticker = "PYTH",
+            logo = "pyth",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "pyth-network",
+            contractAddress = "HZ1JovNiVvGrGNiiYvEozEVgZ58xaU3RKwX8eACQBCt3",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Solana,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "usd-coin",
+            contractAddress = "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Solana,
+            ticker = "USDT",
+            logo = "usdt",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "tether",
+            contractAddress = "Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Solana,
+            ticker = "RAY",
+            logo = "ray",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "raydium",
+            contractAddress = "4k3Dyjzvzp8eMZWUXbBCjEvwSkkk59S5iCNLY3QrkX6R",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "BNB",
+            logo = "bsc",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "binancecoin",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "USDT",
+            logo = "usdt",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "tether",
+            contractAddress = "0x55d398326f99059fF775485246999027B3197955",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "usd-coin",
+            contractAddress = "0x8ac76a51cc950d9822d68b83fe1ad97b32cd580d",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "DAI",
+            logo = "dai",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "ETH",
+            logo = "eth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x2170ed0880ac9a755fd29b2688956bd959f933f8",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "AAVE",
+            logo = "aave",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xfb6115445bff7b52feb98650c87f44907e58f802",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "COMP",
+            logo = "comp",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x52ce071bd9b1c4b00a0b92d298c512478cad67e8",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "SUSHI",
+            logo = "sushi",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x947950bcc74888a40ffa2593c5798f11fc9124c4",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "KNC",
+            logo = "knc",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xfe56d5892bdffc7bf58f2e84be1b2c32d21c308b",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = BscChain,
+            ticker = "PEPE",
+            logo = "pepe",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x25d887ce7a35172c62febfd67a1856f20faebb00",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "ETH",
+            logo = "eth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "usd-coin",
+            contractAddress = "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "DAI",
+            logo = "dai",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "dai",
+            contractAddress = "0x50c5725949A6F0c72E6C4a641F24049A917DB0Cb",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "rETH",
+            logo = "reth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "reth",
+            contractAddress = "0xB6fe221Fe9EeF5aBa221c348bA20A1Bf5e73624c",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "ezETH",
+            logo = "ezeth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ezETH",
+            contractAddress = "0x2416092f143378750bb29b79eD961ab195CcEea5",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "PYTH",
+            logo = "pyth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "pyth",
+            contractAddress = "0x4c5d8A75F3762c1561D96f177694f67378705E98",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "OM",
+            logo = "om",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "om",
+            contractAddress = "0x3992B27dA26848C2b19CeA6Fd25ad5568B68AB98",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "W",
+            logo = "w",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "w",
+            contractAddress = "0xB0fFa8000886e57F86dd5264b9582b2Ad87b2b91",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "cbETH",
+            logo = "cbeth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "cbETH",
+            contractAddress = "0x2Ae3F1Ec7F1F5012CFEab0185bfc7aa3cf0DEc22",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "SNX",
+            logo = "snx",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "SNX",
+            contractAddress = "0x22e6966B799c4D5B13BE962E1D117b56327FDa66",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Base,
+            ticker = "AERO",
+            logo = "aero",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "AERO",
+            contractAddress = "0x940181a94A35A4569E4529A3CDfB74e38FD98631",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "ETH",
+            logo = "eth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "ARB",
+            logo = "arbitrum",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "arbitrum",
+            contractAddress = "0x912CE59144191C1204E64559FE8253a0e49E6548",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "TGT",
+            logo = "tgt",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "thorwallet",
+            contractAddress = "0x429fEd88f10285E61b12BDF00848315fbDfCC341",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "FOX",
+            logo = "fox",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "shapeshift-fox-token",
+            contractAddress = "0xf929de51D91C77E42f5090069E0AD7A09e513c73",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "USDT",
+            logo = "usdt",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "USDT",
+            contractAddress = "0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "USDC.e",
+            logo = "usdc.e",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "USDC.e",
+            contractAddress = "0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "USDC",
+            contractAddress = "0xaf88d065e77c8cC2239327C5EDb3A432268e5831",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "WBTC",
+            logo = "wbtc",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "WBTC",
+            contractAddress = "0x2f2a2543B76A4166549F7aaB2e75Bef0aefC5B0f",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "LINK",
+            logo = "link",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "LINK",
+            contractAddress = "0xf97f4df75117a78c1A5a0DBb814Af92458539FB4",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "DAI",
+            logo = "dai",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "DAI",
+            contractAddress = "0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "UNI",
+            logo = "uni",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "UNI",
+            contractAddress = "0xFa7F8980b0f1E64A2062791cc3b0871572f1F7f0",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "PEPE",
+            logo = "pepe",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "PEPE",
+            contractAddress = "0x25d887Ce7a35172C62FeBFD67a1856F20FaEbB00",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "GRT",
+            logo = "grt",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "GRT",
+            contractAddress = "0x9623063377AD1B27544C965cCd7342f7EA7e88C7",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "ezETH",
+            logo = "ezeth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ezETH",
+            contractAddress = "0x2416092f143378750bb29b79eD961ab195CcEea5",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "PYTH",
+            logo = "pyth",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "PYTH",
+            contractAddress = "0xE4D5c6aE46ADFAF04313081e8C0052A30b6Dd724",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Arbitrum,
+            ticker = "LDO",
+            logo = "ldo",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "LDO",
+            contractAddress = "0x13Ad51ed4F1B7e9Dc168d8a00cB3f4dDD85EfA60",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "ETH",
+            logo = "eth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "OP",
+            logo = "optimism",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "optimism",
+            contractAddress = "0x4200000000000000000000000000000000000042",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "FOX",
+            logo = "fox",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "shapeshift-fox-token",
+            contractAddress = "0xf1a0da3367bc7aa04f8d94ba57b862ff37ced174",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "USDT",
+            logo = "usdt",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x94b008aA00579c1307B0EF2c499aD98a8ce58e58",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x0b2C639c533813f4Aa9D7837CAf62653d097Ff85",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "USDC.e",
+            logo = "USDC.e",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x7F5c764cBc14f9669B88837ca1490cCa17c31607",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "WBTC",
+            logo = "wbtc",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x68f180fcCe6836688e9084f035309E29Bf0A2095",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "LINK",
+            logo = "link",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x350a791Bfc2C21F9Ed5d10980Dad2e2638ffa7f6",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "DAI",
+            logo = "dai",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "ezETH",
+            logo = "ezeth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x2416092f143378750bb29b79eD961ab195CcEea5",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "PYTH",
+            logo = "pyth",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x99C59ACeBFEF3BBFB7129DC90D1a11DB0E91187f",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Optimism,
+            ticker = "LDO",
+            logo = "ldo",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xFdb794692724153d1488CcdBE0C56c252596735F",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "POL",
+            logo = "pol",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "matic-network",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "WETH",
+            logo = "weth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "FOX",
+            logo = "fox",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "shapeshift-fox-token",
+            contractAddress = "0x65a05db8322701724c197af82c9cae41195b0aa8",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "USDT",
+            logo = "usdt",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xc2132D05D31c914a87C6611C10748AEb04B58e8F",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "BNB",
+            logo = "bnb",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x3BA4c387f786bFEE076A58914F5Bd38d668B42c3",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "SOL",
+            logo = "sol",
+            address = "",
+            decimal = 9,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xd93f7E271cB87c23AaA73edC008A79646d1F9912",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x3c499c542cEF5E3811e1192ce70d8cC03d5c3359",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "USDC.e",
+            logo = "USDC.e",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "BUSD",
+            logo = "busd",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xdAb529f40E671A1D4bF91361c21bf9f0C9712ab7",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "WBTC",
+            logo = "wbtc",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x1BFD67037B42Cf73acF2047067bd4F2C47D9BfD6",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "AVAX",
+            logo = "avax",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x2C89bbc92BD86F8075d1DEcc58C7F4E0107f286b",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "SHIB",
+            logo = "shib",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x6f8a06447Ff6FcF75d803135a7de15CE88C1d4ec",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Polygon,
+            ticker = "LINK",
+            logo = "link",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xb0897686c545045aFc77CF20eC7A532E3120E0F1",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "ETH",
+            logo = "eth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "WETH",
+            logo = "weth",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "0x4300000000000000000000000000000000000004",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "WBTC",
+            logo = "wbtc",
+            address = "",
+            decimal = 8,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xF7bc58b8D8f97ADC129cfC4c9f45Ce3C0E1D2692",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "USDB",
+            logo = "usdb",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x4300000000000000000000000000000000000003",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "BLAST",
+            logo = "blast",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xb1a5700fA2358173Fe465e6eA4Ff52E36e88E2ad",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "MIM",
+            logo = "mim",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x76DA31D7C9CbEAE102aff34D3398bC450c8374c1",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "bLOOKS",
+            logo = "blooks",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x406F10d635be12ad33D6B133C6DA89180f5B999e",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "BAG",
+            logo = "bag",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0xb9dfCd4CF589bB8090569cb52FaC1b88Dbe4981F",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "ZERO",
+            logo = "zero",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x357f93E17FdabEcd3fEFc488a2d27dff8065d00f",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "AI",
+            logo = "ai",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x764933fbAd8f5D04Ccd088602096655c2ED9879F",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "JUICE",
+            logo = "juice",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x818a92bc81Aad0053d72ba753fb5Bc3d0C5C0923",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "OMNI",
+            logo = "omni",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x9e20461bc2c4c980f62f1B279D71734207a6A356",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.Blast,
+            ticker = "DACKIE",
+            logo = "dackie",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "",
+            contractAddress = "0x47C337Bd5b9344a6F3D6f58C474D9D8cd419D8cA",
+            isNativeToken = false,
+        ),
+        Coin(
+            chain = Chain.CronosChain,
+            ticker = "CRO",
+            logo = "cro",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "crypto-com-chain",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.GaiaChain,
+            ticker = "ATOM",
+            logo = "atom",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "cosmos",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Kujira,
+            ticker = "KUJI",
+            logo = "kuji",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "kujira",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Dydx,
+            ticker = "DYDX",
+            logo = "dydx",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "dydx-chain",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        polkadot,
+        Coin(
+            chain = Chain.ZkSync,
+            ticker = "ETH",
+            logo = "zksync",
+            address = "",
+            decimal = 18,
+            hexPublicKey = "",
+            priceProviderID = "ethereum",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Sui,
+            ticker = "Sui",
+            logo = "sui",
+            address = "",
+            decimal = 9,
+            hexPublicKey = "",
+            priceProviderID = "sui",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Ton,
+            ticker = "TON",
+            logo = "ton",
+            address = "",
+            decimal = 9,
+            hexPublicKey = "",
+            priceProviderID = "the-open-network",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Osmosis,
+            ticker = "OSMO",
+            logo = "osmo",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "osmosis",
+            contractAddress = "",
+            isNativeToken = true,
+        ),
+        Coin(
+            chain = Chain.Noble,
+            ticker = "USDC",
+            logo = "usdc",
+            address = "",
+            decimal = 6,
+            hexPublicKey = "",
+            priceProviderID = "usd-coin",
+            contractAddress = "",
+            isNativeToken = true,
+        )
+    ) + terraTokens + terraClassicTokens
+
+    private val terraTokens
+        get() = listOf(
+            Coin(
+                chain = Terra,
+                ticker = "LUNA",
+                logo = "luna",
+                decimal = 6,
+                priceProviderID = "terra-luna-2",
+                contractAddress = "",
+                isNativeToken = true,
+                address = "",
+                hexPublicKey = "",
+            ),
+            Coin(
+                chain = Terra,
+                ticker = "TPT",
+                logo = "terra-poker-token",
+                decimal = 6,
+                priceProviderID = "tpt",
+                contractAddress = "terra13j2k5rfkg0qhk58vz63cze0uze4hwswlrfnm0fa4rnyggjyfrcnqcrs5z2",
+                isNativeToken = false,
+                address = "",
+                hexPublicKey = "",
+            ),
+            Coin(
+                chain = Terra,
+                ticker = "ASTRO-IBC",
+                logo = "astro",
+                decimal = 6,
+                priceProviderID = "astroport-fi",
+                contractAddress = "ibc/8D8A7F7253615E5F76CB6252A1E1BD921D5EDB7BBAAF8913FB1C77FF125D9995",
+                isNativeToken = false,
+                address = "",
+                hexPublicKey = "",
+            ),
+            Coin(
+                chain = Terra,
+                ticker = "ASTRO",
+                logo = "astro",
+                decimal = 6,
+                priceProviderID = "astroport-fi",
+                contractAddress = "terra1nsuqsk6kh58ulczatwev87ttq2z6r3pusulg9r24mfj2fvtzd4uq3exn26",
+                isNativeToken = false,
+                address = "",
+                hexPublicKey = "",
+            ),
+        )
+
+    private val terraClassicTokens
+        get() = listOf(
+            Coin(
+                chain = TerraClassic,
+                ticker = "LUNC",
+                logo = "lunc",
+                decimal = 6,
+                priceProviderID = "terra-luna",
+                contractAddress = "",
+                isNativeToken = true,
+                address = "",
+                hexPublicKey = "",
+            ),
+            Coin(
+                chain = TerraClassic,
+                ticker = "USTC",
+                logo = "ustc",
+                decimal = 6,
+                priceProviderID = "terrausd",
+                contractAddress = "uusd",
+                isNativeToken = false,
+                address = "",
+                hexPublicKey = "",
+            ),
+            Coin(
+                chain = TerraClassic,
+                ticker = "ASTROC",
+                logo = "astro",
+                decimal = 6,
+                priceProviderID = "astroport",
+                contractAddress = "terra1xj49zyqrwpv5k928jwfpfy2ha668nwdgkwlrg3",
+                isNativeToken = false,
+                address = "",
+                hexPublicKey = "",
+            )
+        )
+
+}
